add revert to height

+++ cmd/utils/flags.go
@@ -68,7 +68,7 @@ var (
 		Usage: "Block data storage `<path>`",
 		Value: config.DEFAULT_DATA_DIR,
 	}
-	RevertToHeightFlag = cli.IntFlag{
+	RevertToHeightFlag = cli.Uint64Flag{
 		Name:  "height",
 		Usage: "revert current info to height",
 		Value: config.DEFAULT_REVERT_HEIGHT,
